Pull out SkillCommand
